Duplicate methods eliminated
