+++ backend/Origam.Gui/FormTools.cs
@@ -190,7 +190,7 @@ namespace Origam.Gui
 
             DatasetGenerator gen = new DatasetGenerator(true);
             DataSet sdData = gen.CreateDataSet(entity);
-            RemoveNullConstraints(sdData);
+            sdData.RemoveNullConstraints();
 
             if (transformationBeforeId != Guid.Empty)
             {
@@ -217,68 +217,7 @@ namespace Origam.Gui
             }
             return sdData;
         }
-
-        public static void RemoveNullConstraints(DataSet data)
-        {
-            foreach (DataTable table in data.Tables)
-            {
-                foreach (DataColumn col in table.Columns)
-                {
-                    if (col.AllowDBNull == false & IsKey(col) == false)
-                    {
-                        col.AllowDBNull = true;
-                    }
-                }
-            }
-        }
-
-        private static bool IsKey(DataColumn column)
-        {
-            // primary key
-            bool found = IsInColumns(column, column.Table.PrimaryKey);
-            if (found) return true;
-
-            // parent relations
-            found = IsInRelations(column, column.Table.ParentRelations);
-            if (found) return true;
-
-            // child relations
-            return IsInRelations(column, column.Table.ChildRelations);
-        }
-
-        private static bool IsInRelations(DataColumn column, DataRelationCollection relations)
-        {
-            foreach (DataRelation relation in relations)
-            {
-                if (IsRelationKey(column, relation)) return true;
-            }
-
-            return false;
-        }
-
-        private static bool IsRelationKey(DataColumn column, DataRelation relation)
-        {
-            // parent columns
-            bool found = IsInColumns(column, relation.ParentColumns);
-
-            if (found) return true;
-
-            // child columns
-            return IsInColumns(column, relation.ChildColumns);
-        }
-
-
-        private static bool IsInColumns(DataColumn searchedColumn, DataColumn[] columns)
-        {
-            foreach (DataColumn col in columns)
-            {
-                if (col.Equals(searchedColumn)) return true;
-            }
-
-            return false;
-        }
-
-
+        
         public static DataRow GetSelectionDialogResultRow(Guid entityId, Guid transformationAfterId, IDataDocument dataDoc, object profileId)
         {
             IPersistenceService persistence = ServiceManager.Services.GetService(typeof(IPersistenceService)) as IPersistenceService;
